+++ core-lib/.gitignore
