+++ core-lib/pom.xml
