+++ core-lib/src/main/java/ru/abenefic/spring/shop/core/exceptions/ResourceNotFoundException.java
