+++ core-lib/src/main/java/ru/abenefic/spring/shop/core/model/User.java
