+++ core-lib/src/main/java/ru/abenefic/spring/shop/core/model/dtos/ProductItemDto.java
@@ -1,0 +1,14 @@
+package ru.abenefic.spring.shop.core.model.dtos;
+
+import lombok.*;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@RequiredArgsConstructor
+public class ProductItemDto {
+    private long id;
+
+    @NonNull
+    private ProductDto product;
+}
