+++ core-lib/src/main/resources/application.yaml
