+++ ms-product/src/main/java/ru/abenefic/spring/shop/product/controller/ProductController.java
@@ -9,8 +9,8 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.*;
 import ru.abenefic.spring.shop.core.exceptions.NoSuchPageException;
+import ru.abenefic.spring.shop.core.model.dtos.ProductDto;
 import ru.abenefic.spring.shop.product.model.Product;
-import ru.abenefic.spring.shop.product.model.ProductDto;
 import ru.abenefic.spring.shop.product.model.ProductSort;
 import ru.abenefic.spring.shop.product.model.SortDirection;
 import ru.abenefic.spring.shop.product.repository.ProductSpecifications;
