+++ ms-product/src/main/resources/db/migration/V1__init.sql
@@ -81,5 +81,3 @@ values (1),
        (5),
        (5),
        (6);
-    (6);
-    (6);
